Move formatters to cnxepub/formatters.py
